requirements.txt
requests
tzdata
pycryptodome
frida-tools
